WIP - removing sam cli version validator file
