+++ src/shared/sam/cli/samCliInvokerUtils.ts
@@ -7,27 +7,8 @@
 
 import { SpawnOptions } from 'child_process'
 import { ChildProcessResult } from '../../utilities/childProcess'
-import { SamCliVersionValidatorResult } from './samCliVersionValidator'
 
 export interface SamCliProcessInvoker {
     invoke(options: SpawnOptions, ...args: string[]): Promise<ChildProcessResult>
     invoke(...args: string[]): Promise<ChildProcessResult>
-}
-
-export class InvalidSamCliError extends Error {
-    public constructor(message?: string | undefined) {
-        super(message)
-    }
-}
-
-export class SamCliNotFoundError extends InvalidSamCliError {
-    public constructor() {
-        super('SAM CLI was not found')
-    }
-}
-
-export class InvalidSamCliVersionError extends InvalidSamCliError {
-    public constructor(public versionValidation: SamCliVersionValidatorResult) {
-        super('SAM CLI has an invalid version')
-    }
 }
